events: extract addAbortListener for safe internal use
Refs: https://github.com/nodejs/node/pull/48596
