+++ lib/events.js
@@ -48,7 +48,6 @@ const {
   Symbol,
   SymbolFor,
   SymbolAsyncIterator,
-  SymbolDispose,
 } = primordials;
 const kRejection = SymbolFor('nodejs.rejection');
 
@@ -84,6 +83,7 @@ const {
   validateObject,
   validateString,
 } = require('internal/validators');
+const { addAbortListener } = require('internal/events/abort_listener');
 
 const kCapture = Symbol('kCapture');
 const kErrorMonitor = Symbol('events.errorMonitor');
@@ -1219,35 +1219,6 @@ function listenersController() {
       while (listeners.length > 0) {
         ReflectApply(eventTargetAgnosticRemoveListener, undefined, ArrayPrototypePop(listeners));
       }
-    },
-  };
-}
-
-let queueMicrotask;
-
-function addAbortListener(signal, listener) {
-  if (signal === undefined) {
-    throw new ERR_INVALID_ARG_TYPE('signal', 'AbortSignal', signal);
-  }
-  validateAbortSignal(signal, 'signal');
-  validateFunction(listener, 'listener');
-
-  let removeEventListener;
-  if (signal.aborted) {
-    queueMicrotask ??= require('internal/process/task_queues').queueMicrotask;
-    queueMicrotask(() => listener());
-  } else {
-    kResistStopPropagation ??= require('internal/event_target').kResistStopPropagation;
-    // TODO(atlowChemi) add { subscription: true } and return directly
-    signal.addEventListener('abort', listener, { __proto__: null, once: true, [kResistStopPropagation]: true });
-    removeEventListener = () => {
-      signal.removeEventListener('abort', listener);
-    };
-  }
-  return {
-    __proto__: null,
-    [SymbolDispose]() {
-      removeEventListener?.();
     },
   };
 }
